backend/requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
boto3==1.35.24
botocore==1.35.24
PyMySQL==1.1.1
python-dotenv==1.0.1
pytest==8.3.3
